Do not overwrite subdissector infos and protocols.
The ZMTP dissector sets the info and proto columns after calling
subdissectors. This means it is not possible for a subdissector to set
it, changes will be overwritten.

This is annoying, because usually the subdissector has more relevant
application info.

Remember when a sub-dissector is found, and in this case, do not set
anything in the proto and info columns at our level.