Allow subdissectors to register for a specific TCP port
Replace the name-based table for subdissectors with a table based on the
TCP port. This allows using multiple subdissectors in the same capture.

Leave the responsibility of registering TCP ports to subdissectors
(unfortunately it does not seem possible to list the ports of our
dissector table and push them into the tcp.port table). Alternatively,
"Decode as" can be used to decode ZMTP on a specific port.

As a result, no preferences and configuration are needed anymore. Update
the README accordingly.

Signed-off-by: Adrien Destugues <adrien.destugues@opensource.viveris.fr>